+++ Engine/Physics/Private/Physics.cpp
@@ -3,7 +3,7 @@
 namespace Engine {
 	namespace Physics {
 		void UpdateAll(std::vector<GameObject::GameObject*>& sceneObjects, float deltaTime) {
-			for (int i = 0; i < sceneObjects.size(); i++) {
+			for (size_t i = 0; i < sceneObjects.size(); i++) {
 				if (sceneObjects[i]->GetPhysicsBody() != nullptr) {
 					sceneObjects[i]->GetPhysicsBody()->PhysicsUpdate(deltaTime);
 				}
